refactor: provider index, quoter
